tests/wpt/tests/paint-timing/first-paint-only/first-paint-only.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test: FP only</title>
</head>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/utils.js"></script>
<div id="main"></div>

<script>
setup({"hide_test_state": true});
async_test(function (t) {
    assert_implements(window.PerformancePaintTiming, "Paint Timing isn't supported.");
    const div = document.createElement("div");
    div.style.width = "100px";
    div.style.height = "100px";
    div.style.backgroundColor = "red";
    div.style.color = "blue";
    document.getElementById("main").appendChild(div);
    assertFirstPaint(t, 20);

}, "Performance first paint timing entry exists. No first contentful paint.");
</script>
</body>
</html>
