third_party/blink/web_tests/external/wpt/paint-timing/first-paint-only/first-paint-only-expected.txt

This is a testharness.js-based test.
[FAIL] Performance first paint timing entry exists. No first contentful paint.
  assert_equals: FP only. expected 1 but got 0
Harness: the test ran to completion.
